Din v0.1.0 Din

Din is a Discord wrapper for Elixir.

It’s focused around simplicity and makes sure to actually reconnect when a connection or heartbeat drops.

Usage is simple. Include Din in your dependencies and start it from your supervisor.

Example

defmodule YourApplication.YourModule do
  use Din.Module
  alias Din.Resources.Channel

  handle :message_create do
    match "!ping", do: reply "Pong!"
  end

  # Fallback for unused events
  handle_fallback()
end

Types

snowflake()

snowflake() :: String.t

64-bit string IDs

Functions

discord_url()

discord_url() :: String.t

Base url for Discord’s HTTP API