Din v0.1.0 Din.API

REST calls for Discord HTTP API.

Functions

delete(endpoint)

delete(String.t) :: map

DELETE to a given endpoint.

get(endpoint)

get(String.t) :: map

GET from given endpoint.

headers()

headers() :: map

Defines headers for each call.

multipart(endpoint, data)

multipart(String.t, list) :: map

POST a multipart to a given endpoint with supplied keyword list.

parse(map)

parse(map) :: map | Din.Error.t | :ok

Returns Din.Error.t/0 structs, :ok for 204 no response endpoints, or the map of the returned data.

patch(endpoint, data \\ [])

patch(String.t, list) :: map

PATCH to a given endpoint with supplied keyword list.

post(endpoint, data \\ [])

post(String.t, list) :: map

POST to a given endpoint with supplied keyword list.

put(endpoint, data \\ [])

put(String.t, list) :: map

PUT to a given endpoint with supplied keyword list.