Din v0.1.1 Din.Module

Module for pluggable implementation for any application.

Functions

__using__(opts) (macro)

__using__(list) :: any

Creates a GenServer for your handler file.

Just add use Din.Module at the top of your handler, and start it using YourApplication.YourModule.start_link.

enforce(validator, list) (macro)

enforce(atom, [{:do, any}]) :: any

Macro to create enforcement validators.

Simply a readable version of what will expand to an if statement. The atom should be the name of a function that takes data as a single argument, and should evaluate to either true or false.

Example

alias Din.Resources.Channel

handle :message_create do
  enforce :direct_message do
    match "hello", do: reply "hi!"
  end
end

def direct_message(data) do
  Channel.get(data.channel_id).is_private
end

handle(event, list) (macro)

handle(atom | String.t, [{:do, any}]) :: {:noreply, map}

Macro to create initial handlers for Discord events.

Events can be written as strings or atoms, whichever you prefer. Data sent by the Gateway can be interfaced with the data variable.

Example

handle :message_create do
  IO.inspect data.content
end

handle "PRESENCE_UPDATE" do
  IO.inspect data
end

handle_any(list) (macro)

handle_any([{:do, any}]) :: {:noreply, map}

Macro to handle any event.

Only use this once at the end of your handler module. Data sent by the Gateway can be interfaced with the data variable.

handle_fallback() (macro)

handle_fallback() :: {:noreply, map}

Macro to handle unused events.

Place this once at the end of your handler module if you do not use handle_any/1.

match(text, body) (macro)

match(String.t, [{:do, any}]) :: any

match([String.t], [{:do, any}]) :: any

match(String.t, atom) :: any

match([String.t], atom) :: any

Macro to match text content, as either a string or a list.

Typically used for the MESSAGE_CREATE event. Matches given text with the beginning of the content sent by users using Regex.

The second argument can either be a do block, or an atom of a function that takes data as an argument.

Example

match "!ping", do: IO.inspect data.content
match ["!foo", "!bar"], :inspect

...

def inspect(data) do
  IO.inspect data
end

reply(text, opts \\ []) (macro)

reply(String.t, list) :: any

Macro to reply to messages.

Typically used for the MESSAGE_CREATE event, or anything with a channel_id key. Will accept the same content as Din.Resources.Channel.create_message/3.

Example

match "!ping", do: reply "Pong!"