Din v0.1.1 Din.Resources.Channel

Functions

add_pinned_message(channel_id, message_id)

add_pinned_message(Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Pin a message in a channel.

Requires the MANAGE_MESSAGES permission. Returns :ok on success.

bulk_delete_messages(channel_id, message_ids)

bulk_delete_messages(Din.snowflake, [Din.snowflake]) ::
  :ok |
  Din.Error.t

Delete multiple messages in a single request.

This endpoint can only be used on guild channels and requires the MANAGE_MESSAGES permission. Returns :ok on success. Fires multiple Message Delete Gateway events.

Any message IDs given that do not exist or are invalid will count towards the minimum and maximum message count (currently 2 and 100 respectively). Additionally, duplicated IDs will only be counted once.

This endpoint will not delete messages older than 2 weeks, and will fail if any message provided is older than that. An endpoint will be added in the future to prune messages older than 2 weeks from a channel.

create_invite(channel_id, opts \\ [])

create_invite(Din.snowflake, max_age: integer, max_uses: integer, temporary: boolean, unique: boolean) ::
  map |
  Din.Error.t

Create a new invite object for the channel.

Only usable for guild channels. Requires the CREATE_INSTANT_INVITE permission.

Parameters

max_age - Duration of invite in seconds before expiry, or 0 for never. Defaults to 86400 (24 hours.

max_uses - Max number of uses or 0 for unlimited. Defaults to 0

temporary - Whether this invite only grants temporary membership. Defaults to false.

unique - If true, don’t try to reuse a similar invite (useful for creating many unique one time use invites). Defaults to false.

create_message(channel_id, content, opts \\ [])

create_message(Din.snowflake, String.t, nonce: Din.snowflake, tts: boolean, file: binary, embed: map) ::
  map |
  Din.Error.t

Post a message to a guild text or DM channel.

If operating on a guild channel, this endpoint requires the SEND_MESSAGES permission to be present on the current user. Returns a message object. Fires a Message Create Gateway event.

Parameters

nonce - a nonce that can be used for optimistic message sending

tts - true if this is a TTS message

file - the contents of the file being sent

embed - embedded rich content

create_reaction(channel_id, message_id, emoji)

create_reaction(Din.snowflake, Din.snowflake, String.t) ::
  :ok |
  Din.Error.t

Create a reaction for the message.

This endpoint requires the READ_MESSAGE_HISTORY permission to be present on the current user. Additionally, if nobody else has reacted to the message using this emoji, this endpoint requires the ADD_REACTIONS permission to be present on the current user. Returns :ok on success.

delete(channel_id)

delete(Din.snowflake) :: map | Din.Error.t

Delete a channel, or close a private message.

Requires the MANAGE_CHANNELS permission for the guild. Deleting a category does not delete its child channels; they will have their parent_id removed and a Channel Update Gateway event will fire for each of them. Returns a channel object on success. Fires a Channel Delete Gateway event.

Deleting a guild channel cannot be undone. Use this with caution, as it is impossible to undo this action when performed on a guild channel. In contrast, when used with a private message, it is possible to undo the action by opening a private message with the recipient again.

delete_all_reactions(channel_id, message_id)

delete_all_reactions(Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Deletes all reactions on a message.

This endpoint requires the MANAGE_MESSAGES permission to be present on the current user.

delete_message(channel_id, message_id)

delete_message(Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Delete a message.

If operating on a guild channel and trying to delete a message that was not sent by the current user, this endpoint requires the MANAGE_MESSAGES permission. Returns a :ok on success. Fires a Message Delete Gateway event.

delete_permission(channel_id, overwrite_id)

delete_permission(Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Delete a channel permission overwrite for a user or role in a channel.

Only usable for guild channels. Requires the MANAGE_ROLES permission. Returns :ok on success.

delete_pinned_message(channel_id, message_id)

delete_pinned_message(Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Delete a pinned message in a channel.

Requires the MANAGE_MESSAGES permission. Returns :ok on success.

delete_reaction(channel_id, message_id, emoji)

delete_reaction(Din.snowflake, Din.snowflake, String.t) ::
  :ok |
  Din.Error.t

Delete a reaction the current user has made for the message.

Returns :ok on success.

delete_user_reaction(channel_id, message_id, emoji, user_id)

delete_user_reaction(Din.snowflake, Din.snowflake, String.t, Din.snowflake) ::
  :ok |
  Din.Error.t

Deletes another user’s reaction.

This endpoint requires the MANAGE_MESSAGES permission to be present on the current user. Returns :ok on success.

edit_message(channel_id, message_id, opts \\ [])

edit_message(Din.snowflake, Din.snowflake, content: String.t, embed: map) ::
  map |
  Din.Error.t

Edit a previously sent message.

You can only edit messages that have been sent by the current user. Returns a message object. Fires a Message Update Gateway event.

Parameters

content - the new message contents (up to 2000 characters)

embed - embedded rich content

edit_permissions(channel_id, overwrite_id, opts)

edit_permissions(Din.snowflake, Din.snowflake, allow: integer, deny: integer, type: String.t) ::
  :ok |
  Din.Error.t

Edit the channel permission overwrites for a user or role in a channel.

Only usable for guild channels. Requires the MANAGE_ROLES permission. Returns :ok on success.

get(channel_id)

get(Din.snowflake) :: map | Din.Error.t

Get a channel by ID.

Returns a channel object.

get_invites(channel_id)

get_invites(Din.snowflake) :: [map] | Din.Error.t

Returns a list of invite objects (with invite metadata) for the channel.

Only usable for guild channels. Requires the MANAGE_CHANNELS permission.

get_message(channel_id, message_id)

get_message(Din.snowflake, Din.snowflake) :: map | Din.Error.t

Returns a specific message in the channel.

If operating on a guild channel, this endpoints requires the READ_MESSAGE_HISTORY permission to be present on the current user. Returns a message object on success.

get_messages(channel_id, opts \\ [])

get_messages(Din.snowflake, around: Din.snowflake, before: Din.snowflake, after: Din.snowflake, limit: 1..100) ::
  [map] |
  Din.Error.t

Returns the messages for a channel.

If operating on a guild channel, this endpoint requires the READ_MESSAGES permission to be present on the current user. Returns an array of message objects on success.

Parameters

around - get messages around this message ID

before - get messages before this message ID

after - get messages after this message ID

limit - max number of messages to return (1-100)

get_pinned_messages(channel_id)

get_pinned_messages(Din.snowflake) :: [map] | Din.Error.t

Returns all pinned messages in the channel as an array of message objects.

get_reactions(channel_id, message_id, emoji)

get_reactions(Din.snowflake, Din.snowflake, String.t) ::
  [map] |
  Din.Error.t

Get a list of users that reacted with this emoji.

Returns an array of user objects on success.

group_dm_add_recipient(channel_id, user_id, opts \\ [])

group_dm_add_recipient(Din.snowflake, Din.snowflake, access_token: String.t, nick: String.t) ::
  map |
  Din.Error.t

Adds a recipient to a Group DM using their access token.

Parameters

access_token - access token of a user that has granted your app the gdm.join scope

nick - nickname of the user being added

group_dm_delete_recipient(channel_id, user_id)

group_dm_delete_recipient(Din.snowflake, Din.snowflake) ::
  map |
  Din.Error.t

Removes a recipient from a Group DM.

modify(channel_id, opts \\ [])

modify(Din.snowflake, name: String.t, position: integer, topic: String.t, nsfw: boolean, bitrate: integer, user_limit: integer, permission_overwrites: [map], parent_id: Din.snowflake) ::
  map |
  Din.Error.t

Update a channels settings.

Requires the MANAGE_CHANNELS permission for the guild. Returns a channel on success, and a Din.Error.t/0 on invalid parameters. Fires a Channel Update Gateway event. If modifying a category, individual Channel Update events will fire for each child channel that also changes.

Parameters

All Channels

name - 2-100 character channel name

position - the position of the channel in the left-hand listing

permission_overwrites - channel or category-specific permissions

parent_id - id of the new parent category for a channel

Text Channels

topic - 0-1024 character channel topic

nsfw - if the channel is nsfw

Voice Channels

bitrate - the bitrate (in bits) of the voice channel; 8000 to 96000 (128000 for VIP servers)

user_limit - the user limit of the voice channel; 0 refers to no limit, 1 to 99 refers to a user limit

trigger_typing_indicator(channel_id)

trigger_typing_indicator(Din.snowflake) :: :ok | Din.Error.t

Post a typing indicator for the specified channel.

Generally bots should not implement this route. However, if a bot is responding to a command and expects the computation to take a few seconds, this endpoint may be called to let the user know that the bot is processing their message. Returns :ok on success. Fires a Typing Start Gateway event.