Din v0.1.0 Din.Resources.Emoji

Functions

create_guild_emoji(guild_id, opts \\ [])

create_guild_emoji(Din.snowflake, name: String.t, image: binary, roles: [Din.snowflake]) ::
  map |
  Din.Error.t

Create a new emoji for the guild.

Returns the new emoji object on success. Fires a Guild Emojis Update Gateway event.

Passing the roles field will be ignored unless the application is whitelisted as an emoji provider. For more information and to request whitelisting please contact support@discordapp.com.

Parameters

name - name of the emoji

image - the 128x128 emoji image

roles - roles for which this emoji will be whitelisted

delete_guild_emoji(guild_id, emoji_id)

delete_guild_emoji(Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Delete the given emoji.

Returns :ok on success. Fires a Guild Emojis Update Gateway event.

get_guild_emoji(guild_id, emoji_id)

get_guild_emoji(Din.snowflake, Din.snowflake) ::
  map |
  Din.Error.t

Returns an emoji object for the given guild and emoji IDs.

list_guild_emojis(guild_id)

list_guild_emojis(Din.snowflake) :: [map] | Din.Error.t

Returns a list of emoji objects for the given guild.

modify_guild_emoji(guild_id, emoji_id, opts \\ [])

modify_guild_emoji(Din.snowflake, Din.snowflake, name: String.t, roles: [Din.snowflake]) ::
  map |
  Din.Error.t

Modify the given emoji.

Returns the updated emoji object on success. Fires a Guild Emojis Update Gateway event.

Passing the roles field will be ignored unless the application is whitelisted as an emoji provider. For more information and to request whitelisting please contact support@discordapp.com

Parameters

name - name of the emoji

roles - roles for which this emoji will be whitelisted