Din v0.1.0 Din.Resources.Guild

Functions

add_member(guild_id, user_id, access_token, opts \\ [])

add_member(Din.snowflake, Din.snowflake, String.t, nick: String.t, roles: [Din.snowflake], mute: boolean, deaf: boolean) ::
  map |
  Din.Error.t

Adds a user to the guild.

You must have a valid oauth2 access token for the user with the guilds.join scope. Returns the guild member as the body. Fires a Guild Member Add Gateway event. Requires the bot to have the CREATE_INSTANT_INVITE permission.

Parameters

access_token - an oauth2 access token granted with the guilds.join to the bot’s application for the user you want to add to the guild

nick - value to set users nickname to

roles - array of role ids the member is assigned

mute - if the user is muted

deaf - if the user is deafened

add_member_role(guild_id, user_id, role_id)

add_member_role(Din.snowflake, Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Adds a role to a guild member.

Requires the MANAGE_ROLES permission. Returns :ok on success. Fires a Guild Member Update Gateway event.

begin_prune(guild_id, opts \\ [])

begin_prune(Din.snowflake, [{:days, integer}]) ::
  map |
  Din.Error.t

Begin a prune operation.

Requires the KICK_MEMBERS permission. Returns an object with one pruned key indicating the number of members that were removed in the prune operation. Fires multiple Guild Member Remove Gateway events.

Parameters

days - number of days to count prune for (1 or more)

create(opts \\ [])

create(name: String.t, region: String.t, icon: binary, verification_level: integer, default_message_notifications: integer, roles: [map], channels: [map]) ::
  map |
  Din.Error.t

Create a new guild.

Returns a guild object on success. Fires a Guild Create Gateway event.

By default this endpoint is limited to 10 active guilds. These limits are raised for whitelisted GameBridge applications.

Creating channel categories from this endpoint is not supported.

If roles are specified, the required id field within each role object is an integer placeholder, and will be replaced by the API upon consumption. Its purpose is to allow you to overwrite a role’s permissions in a channel when also passing in channels with the channels array.

Parameters

name - name of the guild (2-100 characters)

region - {voice_region.id} for voice

icon - base64 128x128 jpeg image for the guild icon

verification_level - guild verification level

default_message_notifications - default message notifications setting

roles - new guild roles

channels - new guild’s channels

create_ban(guild_id, user_id, opts \\ [])

create_ban(Din.snowflake, Din.snowflake, [{:"delete-message_days", integer}]) ::
  :ok |
  Din.Error.t

Create a guild ban.

Optionally delete previous messages sent by the banned user. Requires the BAN_MEMBERS permission. Returns :ok on success. Fires a Guild Ban Add Gateway event.

create_channel(guild_id, channel_name, opts \\ [])

create_channel(Din.snowflake, String.t, type: integer, bitrate: integer, user_limit: integer, permission_overwrites: [map], parent_id: Din.snowflake, nsfw: boolean) ::
  map |
  Din.Error.t

Create a new channel object for the guild.

Requires the MANAGE_CHANNELS permission. Returns the new channel object on success. Fires a Channel Create Gateway event.

create_integration(guild_id, opts \\ [])

create_integration(Din.snowflake, type: String.t, id: Din.snowflake) ::
  :ok |
  Din.Error.t

Attach an integration object from the current user to the guild.

Requires the MANAGE_GUILD permission. Returns :ok on success. Fires a Guild Integrations Update Gateway event.

create_role(guild_id, opts \\ [])

create_role(Din.snowflake, name: String.t, permissions: integer, color: integer, hoist: boolean, mentionable: boolean) ::
  map |
  Din.Error.t

Create a new role for the guild.

Requires the MANAGE_ROLES permission. Returns the new role object on success. Fires a Guild Role Create Gateway event.

Parameters

name - name of the role

permissions - bitwise of the enabled/disabled permissions

color - RGB color value

hoist - whether the role should be displayed separately in the sidebar

mentionable - whether the role should be mentionable

delete(guild_id)

delete(Din.snowflake) :: :ok | Din.Error.t

Delete a guild permanently.

User must be owner. Returns :ok on success. Fires a Guild Delete Gateway event.

delete_integration(guild_id, integration_id)

delete_integration(Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Delete the attached integration object for the guild.

Requires the MANAGE_GUILD permission. Returns :ok on success. Fires a Guild Integrations Update Gateway event.

delete_role(guild_id, role_id)

delete_role(Din.snowflake, Din.snowflake) :: :ok | Din.Error.t

Delete a guild role.

Requires the MANAGE_ROLES permission. Returns :ok on success. Fires a Guild Role Delete Gateway event.

get(guild_id)

get(Din.snowflake) :: map | Din.Error.t

Returns the guild object for the given id.

get_bans(guild_id)

get_bans(Din.snowflake) :: [map] | Din.Error.t

Returns a list of ban objects for the users banned from this guild.

Requires the BAN_MEMBERS permission.

get_channels(guild_id)

get_channels(Din.snowflake) :: [map] | Din.Error.t

Returns a list of guild channel objects.

get_embed(guild_id)

get_embed(Din.snowflake) :: map | Din.Error.t

Returns the guild embed object.

Requires the MANAGE_GUILD permission.

get_integrations(guild_id)

get_integrations(Din.snowflake) :: [map] | Din.Error.t

Returns a list of integration objects for the guild.

Requires the MANAGE_GUILD permission.

get_invites(guild_id)

get_invites(Din.snowflake) :: [map] | Din.Error.t

Returns a list of invite objects (with invite metadata) for the guild.

Requires the MANAGE_GUILD permission.

get_member(guild_id, user_id)

get_member(Din.snowflake, Din.snowflake) :: map | Din.Error.t

Returns a guild member object for the specified user.

get_prune_count(guild_id, opts \\ [])

get_prune_count(Din.snowflake, [{:days, integer}]) ::
  map |
  Din.Error.t

Returns an object with one pruned key indicating the number of members that would be removed in a prune operation.

Requires the KICK_MEMBERS permission.

Parameters

days - number of days to count prune for (1 or more)

get_roles(guild_id)

get_roles(Din.snowflake) :: [map] | Din.Error.t

Returns a list of role objects for the guild.

Requires the MANAGE_ROLES permission.

get_voice_regions(guild_id)

get_voice_regions(Din.snowflake) :: [map] | Din.Error.t

Returns a list of voice region objects for the guild.

Unlike the similar Din.Resources.Voice.list_regions/0, this returns VIP servers when the guild is VIP-enabled.

list_members(guild_id, opts \\ [])

list_members(Din.snowflake, limit: integer, after: Din.snowflake) ::
  [map] |
  Din.Error.t

Returns a list of guild member objects that are members of the guild.

Parameters

limit - max number of members to return (1-1000)

after - the highest user id in the previous page

modify(guild_id, opts \\ [])

modify(Din.snowflake, name: String.t, region: String.t, icon: binary, verification_level: integer, default_message_notifications: integer, afk_channel_id: Din.snowflake, afk_timeout: integer, icon: binary, owner_id: Din.snowflake, splash: String.t) ::
  map |
  Din.Error.t

Modify a guild’s settings.

Returns the updated guild object on success. Fires a Guild Update Gateway event.

Parameters

name - name of the guild (2-100 characters)

region - {voice_region.id} for voice

icon - base64 128x128 jpeg image for the guild icon

verification_level - guild verification level

default_message_notifications - default message notifications setting

afk_channel_id - id for afk channel

afk_timeout - afk timeout in seconds

icon - base64 128x128 jpeg image for the guild icon

owner_id - user id to transfer guild ownership to (must be owner)

splash - base64 128x128 jpeg image for the guild splash (VIP only)

modify_channel_positions(guild_id, channel_list)

modify_channel_positions(Din.snowflake, [[id: Din.snowflake, position: integer]]) ::
  :ok |
  Din.Error.t

Modify the positions of a set of channel objects for the guild.

Requires MANAGE_CHANNELS permission. Returns :ok on success. Fires multiple Channel Update Gateway events.

Only channels to be modified are required, with the minimum being a swap between at least two channels.

Parameters

id - channel id

position - sorting position of the channel

modify_current_users_nick(guild_id, nickname)

modify_current_users_nick(Din.snowflake, String.t) ::
  map |
  Din.Error.t

Modifies the nickname of the current user in a guild.

Returns the nickname on success. Fires a Guild Member Update Gateway event.

modify_embed(guild_id, opts \\ [])

modify_embed(Din.snowflake, enabled: boolean, channel_id: Din.snowflake) ::
  map |
  Din.Error.t

Modify a guild embed object for the guild.

Requires the MANAGE_GUILD permission. Returns the updated guild embed object.

Parameters

enabled - if the embed is enabled

channel_id - the embed channel id

modify_integration(guild_id, integration_id, opts \\ [])

modify_integration(Din.snowflake, Din.snowflake, expire_behavior: integer, expire_grace_period: integer, enable_emoticons: boolean) ::
  :ok |
  Din.Error.t

Modify the behavior and settings of a integration object for the guild.

Requires the MANAGE_GUILD permission. Returns :ok on success. Fires a Guild Integrations Update Gateway event.

Parameters

expire_behavior - the behavior when an integration subscription lapses (see the integration object documentation)

expire_grace_period - period (in seconds) where the integration will ignore lapsed subscriptions

enable_emoticons - whether emoticons should be synced for this integration (twitch only currently)

modify_member(guild_id, user_id, opts \\ [])

modify_member(Din.snowflake, Din.snowflake, nick: String.t, roles: [Din.snowflake], mute: boolean, deaf: boolean, channel_id: Din.snowflake) ::
  :ok |
  Din.Error.t

Modify attributes of a guild member.

Returns :ok on success. Fires a Guild Member Update Gateway event.

When moving members to channels, the API user must have permissions to both connect to the channel and have the MOVE_MEMBERS permission.

Parameters

nick - value to set users nickname to

roles - array of role ids the member is assigned

mute - if the user is muted

deaf - if the user is deafened

channel_id - id of channel to move user to (if they are connected to voice)

modify_role(guild_id, role_id, opts \\ [])

modify_role(Din.snowflake, Din.snowflake, name: String.t, permissions: integer, color: integer, hoist: boolean, mentionable: boolean) ::
  map |
  Din.Error.t

Modify a guild role.

Requires the MANAGE_ROLES permission. Returns the updated role on success. Fires a Guild Role Update Gateway event.

modify_role_positions(guild_id, roles_list)

modify_role_positions(Din.snowflake, [[id: Din.snowflake, position: integer]]) ::
  [map] |
  Din.Error.t

Modify the positions of a set of role objects for the guild.

Requires the MANAGE_ROLES permission. Returns a list of all of the guild’s role objects on success. Fires multiple Guild Role Update Gateway events.

remove_ban(guild_id, user_id)

remove_ban(Din.snowflake, Din.snowflake) :: :ok | Din.Error.t

Remove the ban for a user.

Requires the BAN_MEMBERS permissions. Returns :ok on success. Fires a Guild Ban Remove Gateway event.

remove_member(guild_id, user_id)

remove_member(Din.snowflake, Din.snowflake) :: :ok | Din.Error.t

Remove a member from a guild.

Requires KICK_MEMBERS permission. Returns :ok on success. Fires a Guild Member Remove Gateway event.

remove_member_role(guild_id, user_id, role_id)

remove_member_role(Din.snowflake, Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Removes a role from a guild member.

Requires the MANAGE_ROLES permission. Returns :ok on success. Fires a Guild Member Update Gateway event.

sync_integration(guild_id, integration_id)

sync_integration(Din.snowflake, Din.snowflake) ::
  :ok |
  Din.Error.t

Sync an integration.

Requires the MANAGE_GUILD permission. Returns :ok on success.