Din v0.1.0 Din.Resources.User

Functions

create_dm(opts \\ [])

create_dm([{:recipient_id, Din.snowflake}]) :: map | Din.Error.t

Create a new DM channel with a user.

Returns a DM channel object.

Parameters

recipient_id - the recipient to open a DM channel with

create_group_dm(opts \\ [])

create_group_dm(access_tokens: [String.t], nicks: [%{optional(Din.snowflake) => String.t}]) ::
  map |
  Din.Error.t

Create a new group DM channel with multiple users.

Returns a DM channel object.

Parameters

access_tokens - access tokens of users that have granted your app the gdm.join scope

nicks - a dictionary of user ids to their respective nicknames

get(user_id)

get(Din.snowflake) :: map | Din.Error.t

Returns a user object for a given user ID.

get_connections()

get_connections() :: [map] | Din.Error.t

Returns a list of connection objects.

Requires the connections OAuth2 scope.

get_current_user()

get_current_user() :: map | Din.Error.t

Returns the user object of the requester’s account.

For OAuth2, this requires the identify scope, which will return the object without an email, and optionally the email scope, which returns the object with an email.

get_current_user_guilds(opts \\ [])

get_current_user_guilds(before: Din.snowflake, after: Din.snowflake, limit: integer) ::
  [map] |
  Din.Error.t

Returns a list of partial guild objects the current user is a member of.

Requires the guilds OAuth2 scope.

This endpoint returns 100 guilds by default, which is the maximum number of guilds a non-bot user can join. Therefore, pagination is not needed for integrations that need to get a list of users’ guilds.

get_dms()

get_dms() :: [map] | Din.Error.t

Returns a list of DM channel objects.

leave_guild(guild_id)

leave_guild(Din.snowflake) :: :ok | Din.Error.t

Leave a guild.

Returns :ok on success.

modify_current_user(opts \\ [])

modify_current_user(user: String.t, avatar: binary) ::
  map |
  Din.Error.t

Modify the requester’s user account settings.

Returns a user object on success.