Din v0.1.0 Din.Resources.Webhook

Functions

create(channel_id, opts \\ [])

create(Din.snowflake, name: String.t, avatar: binary) ::
  map |
  Din.Error.t

Create a new webhook.

Returns a webhook object on success.

Parameters

name - name of the webhook (2-32 characters)

avatar - image for the default webhook avatar

delete(webhook_id)

delete(Din.snowflake) :: :ok | Din.Error.t

Delete a webhook permanently.

User must be owner. Returns :ok on success.

delete_with_token(webhook_id, webhook_token)

delete_with_token(Din.snowflake, String.t) :: nil | Din.Error.t

Delete a webhook permanently with token.

Same as delete/1, except this call does not require authentication.

execute(webhook_id, webhook_token, opts \\ [])

execute(Din.snowflake, String.t, wait: boolean, content: String.t, username: String.t, avatar_url: String.t, tts: boolean, file: binary, embeds: [map]) ::
  map |
  Din.Error.t

Sends a message via webhook.

You must have one of any content, file, or embeds fields.

Parameters

wait - waits for server confirmation of message send before response, and returns the created message body (defaults to false; when false a message that is not saved does not return an error)

content - the message contents (up to 2000 characters)

username - override the default username of the webhook

avatar_url - override the default avatar of the webhook

tts - true if this is a TTS message

file - the contents of the file being sent

embeds - embedded rich content

execute_github(webhook_id, webhook_token, opts \\ [])

execute_github(Din.snowflake, String.t, [{:wait, boolean}]) ::
  map |
  Din.Error.t

Execute GitHub compatible webook.

Add a new webhook to your GitHub repo (in the repo’s settings), and use this endpoint as the “Payload URL.” You can choose what events your Discord channel receives by choosing the “Let me select individual events” option and selecting individual events for the new webhook you’re configuring.

Parameters

wait - waits for server confirmation of message send before response, and returns the created message body (defaults to false; when false a message that is not saved does not return an error)

execute_slack(webhook_id, webhook_token, opts \\ [])

execute_slack(Din.snowflake, String.t, [{:wait, boolean}]) ::
  map |
  Din.Error.t

Execute Slack compatible webook.

Refer to Slack’s documentation for more information. Discord does not support Slack’s channel, icon_emoji, mrkdwn, or mrkdwn_in properties.

Parameters

wait - waits for server confirmation of message send before response, and returns the created message body (defaults to false; when false a message that is not saved does not return an error)

get(webhook_id)

get(Din.snowflake) :: map | Din.Error.t

Returns the new webhook object for the given id.

get_channel_webhooks(channel_id)

get_channel_webhooks(Din.snowflake) :: [map] | Din.Error.t

Returns a list of channel webhook objects.

get_guild_webhooks(guild_id)

get_guild_webhooks(Din.snowflake) :: [map] | Din.Error.t

Returns a list of guild webhook objects.

get_with_token(webhook_id, webhook_token)

get_with_token(Din.snowflake, String.t) :: map | Din.Error.t

Returns the new webhook object for the given id and token.

Same as get/1, except this call does not require authentication and returns no user in the webhook object.

modify(webhook_id, opts \\ [])

modify(Din.snowflake, name: String.t, avatar: binary, channel_id: Din.snowflake) ::
  map |
  Din.Error.t

Modify a webhook.

Returns the updated webhook object on success.

Parameters

name - the default name of the webhook

avatar - image for the default webhook avatar

channel_id - the new channel id this webhook should be moved to

modify_with_token(webhook_id, webhook_token, opts \\ [])

modify_with_token(Din.snowflake, String.t, name: String.t, avatar: binary) ::
  map |
  Din.Error.t

Modify a webhook with token.

Same as modify/2, except this call does not require authentication, does not accept a channel_id parameter in the body, and does not return a user in the webhook object.